+++ src/main/java/com/roneet/ticketmanager/dao/CommentInterface.java
@@ -1,4 +1,4 @@
-package com.roneet.ticketmanager.repository;
+package com.roneet.ticketmanager.dao;
 
 import com.roneet.ticketmanager.entity.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
