+++ src/main/java/com/roneet/ticketmanager/dao/UserInterface.java
@@ -1,4 +1,4 @@
-package com.roneet.ticketmanager.repository;
+package com.roneet.ticketmanager.dao;
 
 import com.roneet.ticketmanager.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
