+++ src/main/java/com/roneet/ticketmanager/service/UserService.java
@@ -1,7 +1,7 @@
 package com.roneet.ticketmanager.service;
 
 import com.roneet.ticketmanager.entity.User;
-import com.roneet.ticketmanager.repository.UserInterface;
+import com.roneet.ticketmanager.dao.UserInterface;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
